Added Evaluation Service

+++ Chess/Engine/Models/Moves/MoveBase.cs
@@ -20,8 +20,8 @@ namespace Engine.Models.Moves
 
         #region Implementation of IMove
 
-        public short Static { get; set; }
-        public short Difference { get; set; }
+        public int Static { get; set; }
+        public int Difference { get; set; }
         public int Value { get; set; }
         public Piece Piece { get; set; }
         public Square From { get; set; }
